requirements.txt
Django==4.2
djangorestframework==3.14.0
python_dateutil==2.8.2
requests==2.25.1
